Improve the wrong command format message
The message saying there is wrong number of arguments in the command
will now have quotes around the executable when it is a path containing
spaces.

This allows users to copy the command directly and paste it into the
command line to set the parameters correctly. Previously, if the user
did that they ran into the risk of trying to run a partial directory as
the command with the rest of the path as one or more arguments to it,
along with the arguments this program takes.

This change is also a security fix, since it is possible to construct an
executable file without an extension with the same path as a directory
in some OSes, which will be ran by trying to execute the partial
directory printed in the message. This could be exploited if an attacker
manages to add files to arbitrary locations in the filesystem but can't
run them. It would be possible to trick the user into running the
command copied from the error message, running the arbitrary executable.