Add error checking to number parameter parsing
When number parameter strings were parsed to get integers the old atoi
function was used. It was replaced by the strtol function, with error
handling.

The atoi function has several problems and should be generally avoided
in all cases. The problems include the complete lack of error and
security checking, making the program prone to several exploits and
generally unreliable.

For example, if another program, without elevated privileges, for some
reason, could run this program with elevated privileges, privilege
escalation could be trivially achieved by passing a well-crafted string
as a number parameter that would be passed unsanitized to the atoi
function. Other than that, if remotely running this program with
arbitrary arguments is possible, remote code execution would also be
trivially achievable the same way.

The atoi function also has no error checking or signaling, so if the
number string is not valid its return value is undefined, so it could be
a usable value that has nothing to do with the intended value in the
input string. The strtol function now used has error checking and its
errors are handled by exiting the program with an error message.