Add commit hash to comment that references it
The header for the handleNumberInput function in main.c made a reference
to information in the commit message of the commit that added it. The
commit hash is now in the comment for easier reference.

It may be possible to generate a commit that has a hash that starts with
the same characters as the short commit hash in the comment using
brute-force, but it is not worth the trouble to avoid a single
additional commit.