Change the executable name to "cache_simulator"
The specification requires the generated executable file to be named
"cache_simulator". Previously the default name "main" was used. In
Windows (NT) platforms the executable is named "cache_simulator.exe" by
convention, which likely doesn't break the specification, and will
probably not cause issues in the automated testing since it will almost
certainly be performed in a Linux system.