Improve capacity miss detection performance
The function that tests if the cache is full was called each time a
potentially capacity miss happened, even if the cache was already
detected to be full before. Since when the cache is full it will remain
full until the test is finished, once the cache is detected to be full,
the test will now no longer be preformed again.