Remove miss total tracking
The number of misses was tracked along with the number of each type of
miss. This was not needed for the statistics calculations since the
total can be calculated by adding each miss type. Tracking the total of
misses not only had a negative impact on performance as it also had the
potential to cause errors if the total was not updated alongside each
miss type.