Change input number parsing to use strtoul
The input number parsing was using strtol to parse the input number,
instead of strtoul. The strtoul function works with unsigned longs
instead of longs, which is more appropriate since the input numbers are
always positive.

The handleNumberInput function was also renamed to parseNumberInput to
improve clarity. Comments in it were also change to reflect that it
parses the number input.