Make the cache aware if its replacement policy
The cache previously didn't store its replacement policy, which was a
simulator variable, and not a cache property. Now all the cache accesses
occur with a dispatcher function that will call the appropriate access
function.

This makes the cache more encapsulated, which paves the way
for multi-level caches to be implemented by having lower level caches as
properties in a cache structure and miss handling by forwarding the
access to lower levels.