Implement multiple cache level support
The simulator now supports multiple cache levels. The parameters for the
lower cache levels were already being taken from the command-line
arguments, since the previous commit, but they weren't used.

The implementation was trivial, since the cache was already well
encapsulated, so adding lower levels was a matter of simply having the
lower level cache be a member of the higher level cache struct and
passing addresses that missed to it.

This structure is similar to that of a linked-list: the caches are
chained together in order as a recursive data structure.