Make the cache structure externally accessible
The cache structure and its associated functions are now accessible in
users of the Simulator.h header. Previously, the cache structure could
only be used through the simulate function, not allowing arbitrary
accesses that couldn't be put in the addresses array.

This makes the cache simulation more flexible as it allows for the
simulation to be done in addresses that are not predetermined and for
the full content of the cache to be accessed externally.

This makes the Simulator.h header more useful as a library for general
cache simulations, and could be used in memory tracer-based simulators,
for example.