Add a compliance level parameter
A compliance level compilation parameter can now be set in the
CacheSimulator.h file. This parameter can be set to 0, 1, or 2. Level 0
is relaxed, level 1 is strict, and level 2 is very strict. The default
value is 1.

Compliance level 0 allows for additional features to cause the program
to have different behavior than specified for inputs that are valid per
the specification.

Compliance level 1 only allows additional features that won't affect the
expected behavior of the program for valid inputs, so only additional
features that use inputs that would be otherwise invalid are allowed.

Compliance level 2 completely disallows any additional features, so
invalid inputs will cause errors in all cases in strict conformance with
the specification.